backend/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
pydantic==2.12.5
sqlalchemy==2.0.23
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic-settings==2.1.0
email-validator==2.1.0
google-generativeai==0.8.3
